revert: revert: remove unused wasmer_wamr flag from hc_bundle
