+++ crates/diagnostic_tests/Cargo.toml
@@ -14,7 +14,7 @@ chashmap = "2.2"
 colored = "2.1"
 futures = "0.3"
 kitsune_p2p = { version = "^0.4.0-dev.20", path = "../kitsune_p2p/kitsune_p2p" }
-holochain_diagnostics = { path = "../holochain_diagnostics" }
+holochain_diagnostics = { path = "../holochain_diagnostics", default-features = false }
 holochain_trace = { version = "^0.4.0-dev.5", path = "../holochain_trace" }
 tokio = { version = "1.36.0", features = ["full"] }
 tokio-stream = "0.1"
@@ -26,5 +26,7 @@ tui = "0.19"
 workspace = true
 
 [features]
+default = ["wasmer_sys"]
+
 wasmer_sys = ["holochain_diagnostics/wasmer_sys"]
 wasmer_wamr = ["holochain_diagnostics/wasmer_wamr"]
