+++ crates/hc_bundle/Cargo.toml
@@ -24,7 +24,7 @@ path = "src/bin/hc-dna.rs"
 
 # reminder - do not use workspace deps
 [dependencies]
-holochain_wasmer_host = "=0.0.96"
+holochain_wasmer_host = { version = "=0.0.96", default-features = false }
 futures = "0.3"
 anyhow = "1.0"
 clap = { version = "4.0", features = ["derive"] }
@@ -38,7 +38,10 @@ serde_yaml = "0.9"
 thiserror = "1.0.22"
 tracing = "0.1"
 tokio = { version = "1.27", features = ["full"] }
-wasmer = "=4.3.6"
+
+# Temporarily include a fork of wasmer from the git branch 'wamr', until it is officially released in wasmer v5
+hc-wasmer = { version = "=4.3.6-hc.1", optional = true, default-features = false }
+wasmer = { version = "=4.3.6", optional = true, default-feature = false }
 
 [dev-dependencies]
 assert_cmd = "2.0"
@@ -53,5 +56,18 @@ walkdir = "2"
 workspace = true
 
 [features]
+default = ["wasmer_sys"]
+
 sqlite-encrypted = ["holochain_types/sqlite-encrypted"]
 sqlite = ["holochain_types/sqlite"]
+
+wasmer_sys = [
+  "dep:wasmer",
+  "wasmer/default",
+  "holochain_wasmer_host/wasmer_sys",
+]
+wasmer_wamr = [
+  "dep:hc-wasmer",
+  "hc-wasmer/wamr",
+  "holochain_wasmer_host/wasmer_wamr",
+]
